Menubar track info bug fixes, extensions refactoring
