+++ Tuneful/Extensions/Color+Extension.swift
@@ -1,0 +1,18 @@
+//
+//  Color+Extension.swift
+//  Tuneful
+//
+//  Created by Martin Fekete on 11/01/2024.
+//
+
+import SwiftUI
+
+extension Color {
+    static let playbackPositionLeadingRectangle = Color.init(
+        NSColor(named: .init("playbackPositionLeadingRectangle"))!
+    )
+    
+    static let sliderTrailingRectangle = Color.init(
+        NSColor(named: .init("sliderTrailingRectangle"))!
+    )
+}
