+++ Tuneful/Extensions/DateComponentsFormatter+Extension.swift
@@ -1,0 +1,26 @@
+//
+//  DateComponentsFormatter+Extension.swift
+//  Tuneful
+//
+//  Created by Martin Fekete on 11/01/2024.
+//
+
+import Foundation
+
+extension DateComponentsFormatter {
+    static let playbackTimeWithHours: DateComponentsFormatter = {
+        let formatter = DateComponentsFormatter()
+        formatter.allowedUnits = [.hour, .minute, .second]
+        formatter.unitsStyle = .positional
+        formatter.zeroFormattingBehavior = .pad
+        return formatter
+    }()
+    
+    static let playbackTime: DateComponentsFormatter = {
+        let formatter = DateComponentsFormatter()
+        formatter.allowedUnits = [.minute, .second]
+        formatter.unitsStyle = .positional
+        formatter.zeroFormattingBehavior = .pad
+        return formatter
+    }()
+}
