+++ Tuneful/Extensions/Double+Extension.swift
