+++ Tuneful/Extensions/NSImage+Extension.swift
@@ -1,8 +1,8 @@
 //
-//  NSImage+Empty.swift
+//  NSImage+Extension.swift
 //  Tuneful
 //
-//  Source: [Jukebox](https://github.com/Jaysce/Jukebox)
+//  Created by Martin Fekete on 11/01/2024.
 //
 
 import Foundation
@@ -29,6 +29,26 @@ extension NSImage {
             }
         }
         return true
+    }
+    
+    func roundImage(withSize imageSize: NSSize, radius: CGFloat) -> NSImage {
+        let imageFrame = NSRect(origin: .zero, size: imageSize)
+
+        let newImage = NSImage(size: imageSize)
+        newImage.lockFocus()
+        NSGraphicsContext.saveGraphicsState()
+
+        let path = NSBezierPath(roundedRect: imageFrame, xRadius: radius, yRadius: radius)
+        path.addClip()
+
+        self.size = imageFrame.size
+        self.draw(in: imageFrame, from: NSZeroRect, operation: NSCompositingOperation.sourceOver, fraction: 1.0, respectFlipped: true, hints: nil)
+
+        NSGraphicsContext.restoreGraphicsState()
+
+        newImage.unlockFocus()
+
+        return newImage
     }
 }
