+++ Tuneful/Extensions/Notification.Name.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 extension Notification.Name {
-    
     static let spotifyPlayerStateDidChange = Notification.Name(
         "com.spotify.client.PlaybackStateChanged"
     )
