+++ Tuneful/Info.plist
@@ -2,6 +2,8 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>LSUIElement</key>
+	<true/>
 	<key>SUFeedURL</key>
 	<string>https://martinfekete10.github.io/Tuneful/Support/appcast.xml</string>
 	<key>SUPublicEDKey</key>
