+++ Tuneful/Views/AlbumArtView.swift
@@ -25,7 +25,7 @@ struct AlbumArtView: View {
                     .resizable()
                     .scaledToFill()
                     .frame(width: self.imageSize, height: self.imageSize)
-                    .cornerRadius(8)
+                    .cornerRadius(10)
                     .shadow(color: .black.opacity(0.2), radius: 4, x: 0, y: 2)
             }
             .pressButtonStyle()
@@ -56,9 +56,5 @@ struct AlbumArtView: View {
                 self.isShowingPlaybackControls.toggle()
             }
         }
-    }
-    
-    func test() {
-        NSApplication.shared.sendAction(#selector(AppDelegate.showPreferences), to: nil, from: nil)
     }
 }
